Conversion of my config to utilize snacks.nvim and mini.nvim (#2)
* Update init.lua with new plugin specifications

* feat(snacks): enable commonly used Snacks modules (dashboard, picker, lazygit, statuscolumn, quickfile, input, notifier, bigfile)

* Extensive config updates

* Updated commit hashes on plugins

* Update mini.nvim pairs config

Configures to prevent quotes and backticks from adding the pair
character when it's to go back and surround a word. (I still haven't
mastered the surround feature yet.)

* Re-add model selection menu shortcut in CopilotChat

* Minor update to completion

Don't auto-select a completion item, and update key binds to allow for a
few different options for selecting and accepting options

* Added Go/Python Support | Update README

Added Go/Python support with debugging and updated the README with the
help of AI.

* Remove AI Chat submit keybind

Ctrl+Enter does not work in most "terminals", the default Ctrl+s from
CopilotChat.nvim works fine, and is probably best kept as the submit
keybind.

* CopilotChat - Make ctrl+l confirm reset

I use ctrl+hjkl for moving between window splits, but ctrl+l is mapped
in copilotchat to reset and can accidentally be triggered. This adds a
confirmation to that, so I don't accidentally reset my chat.

* Changes to decrease startup time

* Attempt to further decrease startup time

* Add project.nvim

Add project.nvim support back in and make it manual so I can edit files
inside of projects and when I want to include the whole project, I can
use a keymap

* Update colorscheme and add file for personal snippets (LuaSnip)

* Change from CopilotChat.nvim to CodeCompanion

* Updated AI config

* Correct mini.pairs and combine telekasten config

Updates mini.pairs to not pair quotes when surrounding existing
words/letters, also disable backticks in markdown files.
Moved autocommand for telekasten from autocommands.lua into telekasten
config so everything is in one place.

* Correct which-key keybind display

For CodeCompanion which-key is slow to register the keybindings and as a
result on first open of the `g` key in the CodeCompanion window/buffer
it doesn't show all the proper keybinds and makes it difficult to find
what you're looking for. Subsequent opens seem to work fine.

* Optimize NvimTree

Slow to load in some cases. Disabling LSP diagnostics to attempt a
slight performance increase.

* Keymap cleanup

Global keymaps are in keymaps.lua, but specific keymaps have been moved
to their relevant plugin configurations.

* Corrections (Oops!)

Mapping was incorrect, causing errors.

* Cleanup formatting on all config files

Ran confirm.nvim formatting on all files to cleanup indentation, code
formatting, etc.

---------

Co-authored-by: Chris Gatewood <chris.gatewood@talkdesk.com>